chore(engine-dmn): Adjust Test on Internal Context Functions
Description:
- In feel-scala:1.6.2, the evaluation of an internal context function would throw a FeelMissingFunctionException
- In feel-scala:1.7.4, the evaluation will return null instead since dateTime is a context function and should not be resolved. This call will return null.
- The change adapts the test for feel-scala to make this consideration

Related-to: https://github.com/camunda/camunda-bpm-platform/issues/4100
