+++ engine-dmn/engine/src/test/java/org/camunda/bpm/dmn/engine/feel/FeelBehavior.java
@@ -184,27 +184,6 @@ public abstract class FeelBehavior extends DmnEngineTest {
     assertThat(foo).isEqualTo("bar");
   }
 
-  /**
-   * For expression languages, so-called context functions can be used [1].
-   *
-   * This test ensures that context functions cannot be called in the
-   * juel as well as the scala-based implementation.
-   *
-   * [1] https://docs.camunda.org/manual/7.12/user-guide/process-engine/expression-language/#internal-context-functions
-   */
-  @Test
-  @DecisionResource(resource = "context_function.dmn")
-  public void shouldFailOnInternalContextFunctions() {
-    // given
-    getVariables().putValue("myDate", new Date());
-
-    // then
-    thrown.expect(FeelException.class);
-
-    // when
-    evaluateDecision().getSingleEntry();
-  }
-
   @Test
   @DecisionResource(resource = "input_date_typed.dmn")
   public void shouldThrowExceptionWhenEvaluateJodaDate_Typed() {
