+++ engine-dmn/engine/src/test/java/org/camunda/bpm/dmn/engine/feel/ScalaFeelBehaviorTest.java
@@ -16,9 +16,16 @@
  */
 package org.camunda.bpm.dmn.engine.feel;
 
+import static org.assertj.core.api.Assertions.assertThat;
+
+import java.util.Date;
+import org.camunda.bpm.dmn.engine.DmnDecisionResult;
+import org.camunda.bpm.dmn.engine.DmnDecisionResultEntries;
 import org.camunda.bpm.dmn.engine.DmnEngineConfiguration;
 import org.camunda.bpm.dmn.engine.impl.DefaultDmnEngineConfiguration;
+import org.camunda.bpm.dmn.engine.test.DecisionResource;
 import org.camunda.bpm.dmn.feel.impl.scala.ScalaFeelEngineFactory;
+import org.junit.Test;
 
 public class ScalaFeelBehaviorTest extends FeelBehavior {
 
@@ -28,6 +35,30 @@ public class ScalaFeelBehaviorTest extends FeelBehavior {
     configuration.setFeelEngineFactory(new ScalaFeelEngineFactory());
     configuration.init();
     return configuration;
+  }
+
+  /**
+   * For expression languages, so-called context functions can be used [1].
+   *
+   * This test ensures that context functions cannot be called in the
+   * juel as well as the scala-based implementation.
+   *
+   * [1] https://docs.camunda.org/manual/7.12/user-guide/process-engine/expression-language/#internal-context-functions
+   */
+  @Test
+  @DecisionResource(resource = "context_function.dmn")
+  public void shouldReturnNullOnInternalContextFunctions() {
+    // given
+    getVariables().putValue("myDate", new Date());
+
+    // when
+    DmnDecisionResult result = evaluateDecision();
+    DmnDecisionResultEntries firstResult = result.getFirstResult();
+
+    assertThat(firstResult)
+      .hasSize(1)
+      .containsKey(null)
+      .containsValue("foo");
   }
 
 }
